+++ tests/test_occ_site_functions_occupation.py
@@ -391,12 +391,13 @@ def test_occ_fcc_3a(session_shared_datadir):
         "A.down": make_discrete_magnetic_atom(name="A", value=-1, flavor="C"),
         "B.up": make_discrete_magnetic_atom(name="B", value=1, flavor="C"),
         "B.down": make_discrete_magnetic_atom(name="B", value=-1, flavor="C"),
+        "C": xtal.make_atom(name="C"),
     }
     occ_dof = [
-        ["A.up", "A.down"],
-        ["A.up", "A.down", "B.up", "B.down"],
-        ["A.up", "A.down", "B.up", "B.down"],
-        ["A.up", "A.down", "B.up", "B.down"],
+        ["C", "A.up", "A.down"],
+        ["C", "A.up", "A.down", "B.up", "B.down"],
+        ["C", "A.up", "A.down", "B.up", "B.down"],
+        ["C", "A.up", "A.down", "B.up", "B.down"],
     ]
 
     xtal_prim = xtal.Prim(
@@ -439,7 +440,14 @@ def test_occ_fcc_3a(session_shared_datadir):
                 },
             },
             "basis_function_specs": {
-                "dof_specs": {"occ": {"site_basis_functions": "occupation"}}
+                "dof_specs": {
+                    "occ": {
+                        "site_basis_functions": {
+                            "type": "occupation",
+                            "reference_occ": ["C", "C", "C", "C"],
+                        }
+                    }
+                }
             },
         },
         verbose=False,
@@ -450,55 +458,29 @@ def test_occ_fcc_3a(session_shared_datadir):
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0]:
-        # "A.up", "A.down"
+        # "C", "A.up", "A.down"
         expected = np.array(
             [
-                [1.0, 1.0],
-                [0.0, 1.0],
+                [1.0, 1.0, 1.0],
+                [0.0, 1.0, 0.0],
+                [0.0, 0.0, 1.0],
             ]
         )
         assert np.allclose(builder.occ_site_functions[b]["value"], expected)
 
-    for b in [1]:
-        # "A.up", "A.down", "B.up", "B.down"
+    for b in [1, 2, 3]:
+        # "C", "A.up", "A.down", "B.up", "B.down"
         expected = np.array(
             [
-                [1.0, 1.0, 1.0, 1.0],
-                [0.0, 1.0, 0.0, 0.0],
-                [0.0, 0.0, 1.0, 0.0],
-                [0.0, 0.0, 0.0, 1.0],
+                [1.0, 1.0, 1.0, 1.0, 1.0],
+                [0.0, 1.0, 0.0, 0.0, 0.0],
+                [0.0, 0.0, 1.0, 0.0, 0.0],
+                [0.0, 0.0, 0.0, 1.0, 0.0],
+                [0.0, 0.0, 0.0, 0.0, 1.0],
             ]
         )
         assert np.allclose(builder.occ_site_functions[b]["value"], expected)
 
-    for b in [2]:
-        # "A.up", "A.down", "B.up", "B.down"
-        expected = np.array(
-            [
-                [1.0, 1.0, 1.0, 1.0],
-                [0.0, 1.0, 0.0, 0.0],
-                [0.0, 0.0, 1.0, 0.0],
-                [0.0, 0.0, 0.0, 1.0],
-            ]
-        )
-        assert np.allclose(builder.occ_site_functions[b]["value"], expected)
-
-    for b in [3]:
-        # "A.up", "A.down", "B.up", "B.down"
-        expected = np.array(
-            [
-                [1.0, 1.0, 1.0, 1.0],
-                [0.0, 1.0, 0.0, 0.0],
-                [0.0, 0.0, 1.0, 0.0],
-                [0.0, 0.0, 0.0, 1.0],
-            ]
-        )
-        assert np.allclose(builder.occ_site_functions[b]["value"], expected)
-
-    # import os
-    # import pathlib
-    # from utils.helpers import print_expected_cluster_functions_detailed
-    #
     # print_expected_cluster_functions_detailed(
     #     functions,
     #     file=pathlib.Path(os.path.realpath(__file__)).parent
@@ -537,12 +519,13 @@ def test_occ_fcc_3b(session_shared_datadir):
         "A.down": make_discrete_magnetic_atom(name="A", value=-1, flavor="C"),
         "B.up": make_discrete_magnetic_atom(name="B", value=1, flavor="C"),
         "B.down": make_discrete_magnetic_atom(name="B", value=-1, flavor="C"),
+        "C": xtal.make_atom(name="C"),
     }
     occ_dof = [
-        ["A.up", "A.down"],
-        ["A.up", "A.down", "B.up", "B.down"],
-        ["B.up", "B.down", "A.up", "A.down"],
-        ["A.up", "A.down", "B.up", "B.down"],
+        ["A.up", "A.down", "C"],
+        ["A.up", "A.down", "B.up", "B.down", "C"],
+        ["A.up", "A.down", "B.up", "B.down", "C"],
+        ["A.up", "A.down", "B.up", "B.down", "C"],
     ]
 
     xtal_prim = xtal.Prim(
@@ -553,7 +536,6 @@ def test_occ_fcc_3b(session_shared_datadir):
     )
     prim = casmconfig.Prim(xtal_prim)
     assert len(prim.factor_group.elements) == 96
-
     occ_symgroup_rep = prim.occ_symgroup_rep
 
     for i_factor_group, occ_op_rep in enumerate(occ_symgroup_rep):
@@ -586,7 +568,14 @@ def test_occ_fcc_3b(session_shared_datadir):
                 },
             },
             "basis_function_specs": {
-                "dof_specs": {"occ": {"site_basis_functions": "occupation"}}
+                "dof_specs": {
+                    "occ": {
+                        "site_basis_functions": {
+                            "type": "occupation",
+                            "reference_occ": ["C", "C", "C", "C"],
+                        }
+                    }
+                }
             },
         },
         verbose=False,
@@ -597,50 +586,32 @@ def test_occ_fcc_3b(session_shared_datadir):
     print(xtal.pretty_json(_occ_site_functions))
 
     for b in [0]:
-        # "A.up", "A.down"
+        # "A.up", "A.down", "C"
         expected = np.array(
             [
-                [1.0, 1.0],
-                [0.0, 1.0],
+                [1.0, 1.0, 1.0],
+                [1.0, 0.0, 0.0],
+                [0.0, 1.0, 0.0],
             ]
         )
         assert np.allclose(builder.occ_site_functions[b]["value"], expected)
 
-    for b in [1]:
-        # "A.up", "A.down", "B.up", "B.down"
+    for b in [1, 2, 3]:
+        # "A.up", "A.down", "B.up", "B.down", "C"
         expected = np.array(
             [
-                [1.0, 1.0, 1.0, 1.0],
-                [0.0, 1.0, 0.0, 0.0],
-                [0.0, 0.0, 1.0, 0.0],
-                [0.0, 0.0, 0.0, 1.0],
+                [1.0, 1.0, 1.0, 1.0, 1.0],
+                [1.0, 0.0, 0.0, 0.0, 0.0],
+                [0.0, 1.0, 0.0, 0.0, 0.0],
+                [0.0, 0.0, 1.0, 0.0, 0.0],
+                [0.0, 0.0, 0.0, 1.0, 0.0],
             ]
         )
         assert np.allclose(builder.occ_site_functions[b]["value"], expected)
 
-    for b in [2]:
-        # "A.up", "A.down", "B.up", "B.down"
-        expected = np.array(
-            [
-                [1.0, 1.0, 1.0, 1.0],
-                [0.0, 0.0, 0.0, 1.0],
-                [1.0, 0.0, 0.0, 0.0],
-                [0.0, 1.0, 0.0, 0.0],
-            ]
-        )
-        assert np.allclose(builder.occ_site_functions[b]["value"], expected)
-
-    for b in [3]:
-        # "A.up", "A.down", "B.up", "B.down"
-        expected = np.array(
-            [
-                [1.0, 1.0, 1.0, 1.0],
-                [0.0, 1.0, 0.0, 0.0],
-                [0.0, 0.0, 1.0, 0.0],
-                [0.0, 0.0, 0.0, 1.0],
-            ]
-        )
-        assert np.allclose(builder.occ_site_functions[b]["value"], expected)
+    # import os
+    # import pathlib
+    # from utils.helpers import print_expected_cluster_functions_detailed
 
     # print_expected_cluster_functions_detailed(
     #     functions,
@@ -915,7 +886,14 @@ def test_occ_fcc_4b(session_shared_datadir):
                 },
             },
             "basis_function_specs": {
-                "dof_specs": {"occ": {"site_basis_functions": "occupation"}}
+                "dof_specs": {
+                    "occ": {
+                        "site_basis_functions": {
+                            "type": "occupation",
+                            "reference_occ": ["A", "A", "A", "A"],
+                        }
+                    }
+                }
             },
         },
         verbose=False,
@@ -990,4 +968,126 @@ def test_occ_fcc_4b(session_shared_datadir):
     #     / "expected_occ_site_functions_fcc_4.json",
     # )
     with open(session_shared_datadir / "expected_occ_site_functions_fcc_4.json") as f:
+        assert_expected_cluster_functions_detailed(functions, clusters, json.load(f))
+
+
+def test_occ_fcc_5a(session_shared_datadir):
+    """Prim with occupation DoF with Cmagspin A.up, A.down"""
+
+    atom_A_up = xtal.Occupant(
+        name="A",
+        atoms=[
+            xtal.AtomComponent(
+                name="A",
+                coordinate=[0.0, 0.0, 0.0],
+                properties={"Cmagspin": [1.0]},
+            ),
+        ],
+    )
+    atom_A_down = xtal.Occupant(
+        name="A",
+        atoms=[
+            xtal.AtomComponent(
+                name="A",
+                coordinate=[0.0, 0.0, 0.0],
+                properties={"Cmagspin": [-1.0]},
+            ),
+        ],
+    )
+    occupants = {"A.up": atom_A_up, "A.down": atom_A_down}
+    occ_dof = [
+        ["A.up", "A.down"],
+    ]
+
+    xtal_prim = xtal.Prim(
+        lattice=xtal.Lattice(
+            column_vector_matrix=np.array(
+                [
+                    [0.0, 2.0, 2.0],
+                    [2.0, 0.0, 2.0],
+                    [2.0, 2.0, 0.0],
+                ]
+            ),
+        ),
+        coordinate_frac=np.array(
+            [
+                [0.0, 0.0, 0.0],
+            ]
+        ).T,
+        occ_dof=occ_dof,
+        occupants=occupants,
+    )
+    prim = casmconfig.Prim(xtal_prim)
+    assert len(prim.factor_group.elements) == 96
+    occ_symgroup_rep = prim.occ_symgroup_rep
+
+    for i_factor_group, occ_op_rep in enumerate(occ_symgroup_rep):
+        site_rep = prim.integral_site_coordinate_symgroup_rep[i_factor_group]
+        for i_sublat_before, occ_sublat_rep in enumerate(occ_op_rep):
+            site_before = xtal.IntegralSiteCoordinate(i_sublat_before, [0, 0, 0])
+            site_after = site_rep * site_before
+            i_sublat_after = site_after.sublattice()
+            for i_occ_before in range(len(occ_sublat_rep)):
+                i_occ_after = occ_sublat_rep[i_occ_before]
+
+                orientation_name_before = occ_dof[i_sublat_before][i_occ_before]
+                orientation_name_after = occ_dof[i_sublat_after][i_occ_after]
+
+                # assert occupants map (chemical name match)
+                assert (
+                    occupants[orientation_name_before].name()
+                    == occupants[orientation_name_after].name()
+                )
+
+    # print(xtal.pretty_json(xtal_prim.to_dict()))
+
+    builder = build_cluster_functions(
+        prim=xtal_prim,
+        clex_basis_specs={
+            "cluster_specs": {
+                "orbit_branch_specs": {
+                    "2": {"max_length": 4.01},
+                    "3": {"max_length": 1.01},
+                },
+            },
+            "basis_function_specs": {
+                "dof_specs": {"occ": {"site_basis_functions": "chebychev"}}
+            },
+        },
+        verbose=False,
+    )
+    functions, clusters = (builder.functions, builder.clusters)
+
+    _occ_site_functions = builder.occ_site_functions.copy()
+    print(xtal.pretty_json(_occ_site_functions))
+
+    for b in [0]:
+        # "A.up, A.down"
+        expected = np.array(
+            [
+                [1.0, 1.0],
+                [-1.0, 1.0],
+            ]
+        )
+
+        assert np.allclose(
+            get_occ_site_functions(builder.occ_site_functions, b),
+            expected,
+        )
+
+    from casm.bset.json_io import dump
+
+    dump(builder.to_dict(), "cluster_functions.json", force=True)
+
+    # import os
+    # import pathlib
+    # from utils.helpers import print_expected_cluster_functions_detailed
+    #
+    # print_expected_cluster_functions_detailed(
+    #     functions,
+    #     file=pathlib.Path(os.path.realpath(__file__)).parent
+    #     / "data"
+    #     / "expected_occ_site_functions_fcc_5.json",
+    # )
+    with open(session_shared_datadir / "expected_occ_site_functions_fcc_5.json") as f:
         assert_expected_cluster_functions_detailed(functions, clusters, json.load(f))