+++ CHANGELOG.md
@@ -17,12 +17,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `to_dict` methods to `ClusterFunctionsBuilder`, `OrbitMatrixRepBuilder` and `ClusterMatrixRepBuilder`
 - Added `to_dict` and `from_dict` methods to `ExponentSumConstraint`
 - Added `casm.bset.json_io` module
+- Added option to specify occ_site_basis_functions_specs of type occupation with a choice of reference occupant on each sublattice.
+- Added a check that prints an error message and raises an exception if the constant occupation site basis function mixes with other site basis functions.
 
 ### Changed
 
 - Updated `ClexulatorWriter.write` to write a "variables.json.gz" file for each Clexulator (including local Clexulator) which contains the variables used by the jinja2 templates as well as information like basis function formulas generated during the write process
 - Updated `ClexulatorWriter.write` to write a "cluster_functions.json.gz" file with the generated clusters, matrix reps, and functions.
 - Changed the name used for occupation variables. The name "\\phi" is the base, but the number and meaning of the indices can vary depending on the site basis functions. An "occ_site_functions_info" dict in the "variables.json" file contains the string value used as a template ("occ_var_name") and a description of the variable including its indices, if any ("occ_var_desc" and "occ_var_indices").
+
+### Fixed
+
+- Fixed `v1.basic` templates so that the Clexulator compiles for a prim with occupation DoF, but only the constant basis  function is included.
+- Fixed tests with anisotropic occupants and occupation site basis functions.
 
 
 ## [2.0a1] - 2024-08-15